Fix Assessment protobuf type mismatch by consolidating proto definitions
Previously, Assessment-related protobuf messages were duplicated in both
assessments.proto and databricks_tracing.proto, creating two incompatible
Python classes (assessments_pb2.Assessment and databricks_tracing_pb2.Assessment).
This caused a TypeError when TraceInfoV3 (which expects assessments_pb2.Assessment)
received databricks_tracing_pb2.Assessment from entity code.

Changes:
- Updated databricks_tracing.proto to import and reuse AssessmentSource,
  AssessmentError, Expectation, and Feedback from assessments.proto instead
  of duplicating them
- The Assessment message in databricks_tracing.proto now extends the base
  types with an additional trace_location field for Databricks-specific use
- Updated Python entity classes (assessment.py, assessment_error.py,
  assessment_source.py) to import from assessments_pb2 for consistency
- Regenerated protobuf Python files

This ensures all code uses compatible protobuf types and fixes the
test_evaluate_with_static_dataset test.
