+++ mlflow/entities/assessment.py
@@ -12,9 +12,9 @@ from mlflow.entities._mlflow_object import _MlflowObject
 from mlflow.entities.assessment_error import AssessmentError
 from mlflow.entities.assessment_source import AssessmentSource, AssessmentSourceType
 from mlflow.exceptions import MlflowException
-from mlflow.protos.databricks_tracing_pb2 import Assessment as ProtoAssessment
-from mlflow.protos.databricks_tracing_pb2 import Expectation as ProtoExpectation
-from mlflow.protos.databricks_tracing_pb2 import Feedback as ProtoFeedback
+from mlflow.protos.assessments_pb2 import Assessment as ProtoAssessment
+from mlflow.protos.assessments_pb2 import Expectation as ProtoExpectation
+from mlflow.protos.assessments_pb2 import Feedback as ProtoFeedback
 from mlflow.utils.annotations import experimental
 from mlflow.utils.exception_utils import get_stacktrace
 from mlflow.utils.proto_json_utils import proto_timestamp_to_milliseconds
