+++ mlflow/entities/assessment_error.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 
 from mlflow.entities._mlflow_object import _MlflowObject
-from mlflow.protos.databricks_tracing_pb2 import AssessmentError as ProtoAssessmentError
+from mlflow.protos.assessments_pb2 import AssessmentError as ProtoAssessmentError
 
 _STACK_TRACE_TRUNCATION_PREFIX = "[Stack trace is truncated]\n...\n"
 _STACK_TRACE_TRUNCATION_LENGTH = 10000
